+++ SQLSchemaCompare.Infrastructure/SqlScripters/MySqlScripter.cs
@@ -77,13 +77,17 @@ namespace TiCodeX.SQLSchemaCompare.Infrastructure.SqlScripters
             var sb = new StringBuilder();
 
             var targetTable = t.MappedDbObject as ABaseDbTable;
+            if (targetTable == null)
+            {
+                throw new ArgumentException($"{nameof(t.MappedDbObject)} is null");
+            }
 
             // Remove columns
             var columnsToDrop = targetTable.Columns.Where(x => x.MappedDbObject == null).ToList();
             if (columnsToDrop.Count > 0)
             {
                 sb.AppendLine($"ALTER TABLE {this.ScriptHelper.ScriptObjectName(t)}");
-                for (int i = 0; i < columnsToDrop.Count; i++)
+                for (var i = 0; i < columnsToDrop.Count; i++)
                 {
                     sb.Append($"{this.Indent}DROP COLUMN {columnsToDrop[i].Name}");
                     sb.AppendLine(i == columnsToDrop.Count - 1 ? string.Empty : ",");
@@ -105,7 +109,7 @@ namespace TiCodeX.SQLSchemaCompare.Infrastructure.SqlScripters
             if (columnsToAdd.Count > 0)
             {
                 sb.AppendLine($"ALTER TABLE {this.ScriptHelper.ScriptObjectName(t)}");
-                for (int i = 0; i < columnsToAdd.Count; i++)
+                for (var i = 0; i < columnsToAdd.Count; i++)
                 {
                     sb.Append($"{this.Indent}ADD COLUMN {this.ScriptHelper.ScriptColumn(columnsToAdd[i])}");
                     sb.AppendLine(i == columnsToAdd.Count - 1 ? string.Empty : ",");
@@ -224,9 +228,9 @@ namespace TiCodeX.SQLSchemaCompare.Infrastructure.SqlScripters
 
             // If there is a column with descending order, specify the order on all columns
             var scriptOrder = index.ColumnDescending.Any(x => x);
-            var columnList = index.ColumnNames.Select((x, i) => (scriptOrder ?
+            var columnList = index.ColumnNames.Select((x, i) => scriptOrder ?
                 $"{this.ScriptHelper.ScriptObjectName(x)} {(index.ColumnDescending[i] ? "DESC" : "ASC")}" :
-                $"{this.ScriptHelper.ScriptObjectName(x)}"));
+                $"{this.ScriptHelper.ScriptObjectName(x)}");
 
             sb.Append("CREATE ");
             if (indexMySql.IndexType == "FULLTEXT")
